requirements.txt
# ── Core ML / Vision ────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.26.0
einops>=0.7.0

# ── Qwen2.5-VL specific ──────────────────────────────────────────────────────
qwen-vl-utils>=0.0.8

# ── Image processing ─────────────────────────────────────────────────────────
Pillow>=10.0.0

# ── Data & evaluation ────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ── Plotting ─────────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Progress bar ─────────────────────────────────────────────────────────────
tqdm>=4.65.0

# ── Tokenizer utilities ───────────────────────────────────────────────────────
sentencepiece>=0.1.99
protobuf>=3.20.0
